Refactor TestHelpers
