+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.TestCore/Builders/Command/ChargeInformationCommandReceivedEventBuilder.cs
@@ -14,6 +14,7 @@
 
 using GreenEnergyHub.Charges.Domain.Dtos.ChargeInformationCommandReceivedEvents;
 using GreenEnergyHub.Charges.Domain.Dtos.ChargeInformationCommands;
+using GreenEnergyHub.Charges.TestCore.TestHelpers;
 
 namespace GreenEnergyHub.Charges.TestCore.Builders.Command
 {
