+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.TestCore/Builders/Command/OutboxMessageBuilder.cs
@@ -14,6 +14,7 @@
 
 using System;
 using GreenEnergyHub.Charges.Infrastructure.Outbox;
+using GreenEnergyHub.Charges.TestCore.TestHelpers;
 using NodaTime;
 
 namespace GreenEnergyHub.Charges.TestCore.Builders.Command
