+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.Tests/Infrastructure/CimDeserialization/ChargeBundle/ChargeCommandBundleConverterTests.cs
@@ -27,8 +27,8 @@ using GreenEnergyHub.Charges.Domain.Dtos.ChargeInformationCommands;
 using GreenEnergyHub.Charges.Domain.Dtos.ChargePriceCommands;
 using GreenEnergyHub.Charges.Domain.Dtos.SharedDtos;
 using GreenEnergyHub.Charges.Infrastructure.CimDeserialization.ChargeBundle;
-using GreenEnergyHub.Charges.TestCore;
 using GreenEnergyHub.Charges.TestCore.Attributes;
+using GreenEnergyHub.Charges.TestCore.TestHelpers;
 using GreenEnergyHub.Iso8601;
 using Moq;
 using NodaTime;
