+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.Tests/Infrastructure/CimDeserialization/ChargeBundle/ChargeCommandDeserializerTests.cs
@@ -19,8 +19,8 @@ using System.Threading.Tasks;
 using AutoFixture.Xunit2;
 using Energinet.DataHub.Core.SchemaValidation;
 using GreenEnergyHub.Charges.Infrastructure.CimDeserialization.ChargeBundle;
-using GreenEnergyHub.Charges.TestCore;
 using GreenEnergyHub.Charges.TestCore.Attributes;
+using GreenEnergyHub.Charges.TestCore.TestHelpers;
 using Moq;
 using Xunit;
 using Xunit.Categories;
